Added Post for CIF

+++ Src/Trials/ConsoleApp/BL/Objects/CIFTYPE.cs
@@ -12,7 +12,7 @@ namespace SIS.OpenCore.BL.Objects
 {
     public partial class CifTYPE
     {
-        public static bool ValidateExists(byte TypeCode)
+        public static bool ValidateExists(short TypeCode)
         {
             OpenCoreContext db = new OpenCoreContext();
             var Ret =   (from   t in db.LUT_CIF_TYPE
